+++ src/teams/teams.service.ts
@@ -2,6 +2,8 @@ import { Injectable, NotFoundException } from "@nestjs/common";
 import { UpdateTeamMemberDto } from "./dto/update-team-member.dto";
 import { PrismaService } from "../prisma/prisma.service";
 import { publicVoyageTeamUserSelect } from "../global/selects/teams.select";
+import { UserReq } from "../global/types/CustomRequest";
+import { manageOwnVoyageTeamWithIdParam } from "../ability/conditions/voyage-teams.ability";
 
 @Injectable()
 export class TeamsService {
@@ -25,7 +27,7 @@ export class TeamsService {
         });
     }
 
-    async findTeamById(id: number) {
+    async findTeamById(id: number, user: UserReq) {
         const voyageTeam = await this.prisma.voyageTeam.findUnique({
             where: { id },
             select: publicVoyageTeamUserSelect,
@@ -34,6 +36,7 @@ export class TeamsService {
             throw new NotFoundException(
                 `Voyage team (teamId: ${id}) does not exist.`,
             );
+        manageOwnVoyageTeamWithIdParam(user, id);
         return voyageTeam;
     }